requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
langchain==0.3.27
langchain-google-genai==2.1.12
bcrypt==4.1.2
pydantic[email]==2.7.4
python-multipart==0.0.9
cloudinary==1.40.0
apscheduler==3.10.4
Pillow==10.4.0
pypdf==5.4.0
python-docx==1.1.2
openpyxl==3.1.5
exifread==3.0.0
bleach==6.2.0
beautifulsoup4==4.12.3
markdown-it-py==3.0.0
